refactor: Login 관련 처리를 LoginHandler가 전담하도록 수정 - POST /user/login: 로그인 시도 (세션 생성) - POST /user/logout: 로그아웃 시도 (세션 제거)
