+++ src/main/java/codesquad/application/UserHandler.java
@@ -7,8 +7,6 @@ import codesquad.webserver.handler.SimpleTemplateEngine;
 import codesquad.webserver.http.HttpRequest;
 import codesquad.webserver.http.HttpResponse;
 import codesquad.webserver.http.type.*;
-import codesquad.webserver.session.Session;
-import codesquad.webserver.session.SessionManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -17,7 +15,6 @@ import java.util.Map;
 
 public class UserHandler {
     private static final Logger logger = LoggerFactory.getLogger(UserHandler.class);
-    private static final SessionManager sessionManager = new SessionManager();
 
     private UserHandler() {}
 
@@ -71,15 +68,9 @@ public class UserHandler {
         StringBuilder sb = new StringBuilder();
         users.forEach(user -> {
             sb.append("<tr>");
-            sb.append("<td>");
-            sb.append(user.getName());
-            sb.append("</td>");
-            sb.append("<td>");
-            sb.append(user.getEmail());
-            sb.append("</td>");
-            sb.append("<td>");
-            sb.append(user.getNickname());
-            sb.append("</td>");
+            sb.append("<td>").append(user.getName()).append("</td>");
+            sb.append("<td>").append(user.getEmail()).append("</td>");
+            sb.append("<td>").append(user.getNickname()).append("</td>");
             sb.append("</tr>");
         });
 
@@ -110,44 +101,5 @@ public class UserHandler {
 
         HttpHeader headers = HttpHeader.createRedirection("/index.html");
         return new HttpResponse(HttpProtocol.HTTP_1_1, HttpStatus.FOUND, headers, "유저가 생성되었습니다.");
-    }
-
-    // POST /user/login
-    public static HttpResponse login(HttpRequest httpRequest) {
-        String name = httpRequest.body().get("username");
-        String password = httpRequest.body().get("password");
-
-        try {
-            User user = UserDatabase.getUserByIdAndPassword(name, password);
-            Session session = sessionManager.createSession(user);  // TODO SessionManager 제거
-
-            HttpHeader headers = HttpHeader.createRedirection("/main/index.html");
-            headers.setCookie(Cookie.from(session));
-            return new HttpResponse(HttpProtocol.HTTP_1_1, HttpStatus.FOUND, headers, "로그인 완료!");
-        } catch (IllegalArgumentException e) {
-            logger.debug("유저 정보가 올바르지 않습니다. 이름:{}", name);
-
-            HttpHeader headers = HttpHeader.createRedirection("/user/login_failed.html");
-            return new HttpResponse(HttpProtocol.HTTP_1_1, HttpStatus.FOUND, headers, null);
-        }
-    }
-
-    // POST /user/logout
-    public static HttpResponse logout(HttpRequest httpRequest) {
-        final Cookie cookies = httpRequest.headers().getCookies();
-        final String sid = cookies.get("SID");
-
-        User context = AuthenticationHolder.getContext();
-        if (context == null) {
-            logger.debug("세션이 존재하지 않습니다. sid:{}", sid);
-            return new HttpResponse(HttpProtocol.HTTP_1_1, HttpStatus.UNAUTHORIZED, HttpHeader.createEmpty(), "세션이 존재하지 않습니다.");
-        }
-
-        sessionManager.removeSession(sid);  // TODO SessionManager 제거?
-        logger.debug("정상적으로 로그아웃하였습니다. sid:{}", sid);
-
-        HttpHeader headers = HttpHeader.createRedirection("/index.html");
-        headers.setCookie(Cookie.expiredSession());
-        return new HttpResponse(HttpProtocol.HTTP_1_1, HttpStatus.FOUND, headers, "로그아웃하였습니다.");
     }
 }
